initial code for plant disease detection (ok)

+++ app/src/main/java/com/tiagoleite/detection/Classifier.java
@@ -4,5 +4,5 @@ package com.tiagoleite.detection;
 public interface Classifier
 {
     String name();
-    Classification recognize(final byte pixels[], int w, int h);
+    Classification recognize(final float pixels[], int w, int h);
 }
